Remove NGC5533_functions library First stage of library condensing is complete!
